Rename Android package to match iOS
