+++ android/src/main/AndroidManifest.xml
@@ -1,4 +1,4 @@
-<manifest package="com.reactlibrary"
+<manifest package="com.RNAppleAuthentication"
   xmlns:android="http://schemas.android.com/apk/res/android"
 	xmlns:tools="http://schemas.android.com/tools">
 </manifest>
