+++ android/src/main/java/com/RNAppleAuthentication/SignInWithAppleResult.kt
@@ -1,4 +1,4 @@
-package com.reactlibrary
+package com.RNAppleAuthentication
 
 sealed class SignInWithAppleResult {
   data class Success(val code: String, val id_token: String, val state: String, val user: String) : SignInWithAppleResult()
