+++ example/android/app/src/main/java/com/example/reactnativeappleauth/MainApplication.java
@@ -11,7 +11,7 @@ import com.facebook.soloader.SoLoader;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
 
-import com.reactlibrary.AppleAuthenticationAndroidPackage;
+import com.RNAppleAuthentication.AppleAuthenticationAndroidPackage;
 
 public class MainApplication extends Application implements ReactApplication {
